Preethi/gen-ai-fusion-health-assistant/backend/requirements.txt

fastapi
uvicorn[standard]
pydantic>=2
python-multipart
pillow
numpy
scikit-learn
